+++ README.md
@@ -1,6 +1,6 @@
-# Parish Council Calendar
+# Calendar Scrapers
 
-This repository contains a Python script that scrapes upcoming meeting dates from the St Mewan Parish Council website and builds an iCalendar (`.ics`) file. The generated calendar can be imported into any calendar application that supports the iCalendar format.
+This repository provides small Python scripts that fetch event dates from web pages and build iCalendar (`.ics`) files. The calendars can be imported into any application that supports the iCalendar format.
 
 ## Requirements
 
@@ -16,27 +16,24 @@ pip install -r requirements.txt
 
 ## Usage
 
-Run the script to fetch the meeting information and generate `stmewan.ics`:
+### Penrice Academy Term Dates
+
+Run the script to download the term dates from the Penrice Academy website and generate `penrice.ics`:
 
 ```bash
 python generate_ics.py
 ```
 
-The script downloads the meeting pages, extracts upcoming dates and associated links (agenda and minutes when available) and writes them into `stmewan.ics`.
+### St Mewan Parish Council Meetings
+
+The original scraper for St Mewan Parish Council is kept as `stmewan_generate_ics.py` and produces `stmewan.ics`:
+
+```bash
+python stmewan_generate_ics.py
+```
 
 ## How it works
 
-The `generate_ics.py` script contains a list of meeting pages under `MEETING_TYPES`.
-For each page it:
+Each script downloads the relevant web page, parses the event information using **BeautifulSoup** and writes one calendar entry per event using standard iCalendar syntax.
 
-1. Fetches the HTML with `requests`.
-2. Parses upcoming meeting entries using **BeautifulSoup**.
-3. Converts the textual dates and times into Python `datetime` objects.
-4. Creates an iCalendar event for each meeting, appending any agenda or minutes
-   links to the event description.
-5. Writes all events to `stmewan.ics` in standard iCalendar format.
-
-## Calendar file
-
-An example `stmewan.ics` is included in this repository. After running the script you can import the file into your preferred calendar application to view upcoming Parish Council meetings.
-
+Example calendar files (`penrice.ics` and `stmewan.ics`) are included for reference.
